sub-menu drop down working, styling out of wack
When i did the input[type=checkbox]:checked part for the sub menu, I decided to keep that bit with the sub menu css. THIS WAS A MISTAKE. it ended up lower in the order of priority that sub-menu{display:none} so it would NEVER change the display to block, it would always be of lesser priority than display:none. So the checkbox has been working for the past 2 hours. I am simply an idiot :D